Add demanded features
The sandwich JSON data is parsed by the Utils class to create the sandwich objects. These objects are then used to populate the detail activity. The sandwich origin and additional names are shown based on their availability. To navigate back to the sandwich list, the back button is shown in the action bar.